refactor: add convert method to StatisticsUtils
that method convert Statistics Dto to Map
