--- src/main/java/com/example/spinlog/utils/StatisticsValues.java
@@ -1,5 +1,0 @@
-package com.example.spinlog.utils;
-
-public class StatisticsValues {
-    public static final int PERIOD_CRITERIA = 30;
-}
